cast memory size constant when comparing.

+++ rosmon_core/src/diagnostics/rosmon_to_diagnostic.cpp
@@ -78,21 +78,21 @@ void RosmonToDiagnostic::updateDiagnostics(const std::vector<NodeMonitor::Ptr>& 
 
 std::string RosmonToDiagnostic::memoryToString(uint64_t memory)
 {
-    if(memory < 1000)
+    if(memory < static_cast<uint64_t>(1000))
     {
         return fmt::format("{} B", memory);
     }
-    else if(memory < 1e6)
+    else if(memory < static_cast<uint64_t>(1e6))
     {
-        return fmt::format("{:.2f} KiB", memory / 1e3);
+        return fmt::format("{:.2f} KiB", memory / static_cast<uint64_t>(1e3));
     }
-    else if(memory < 1e9)
+    else if(memory < static_cast<uint64_t>(1e9))
     {
-        return fmt::format("{:.2f} MiB", memory / 1e6);
+        return fmt::format("{:.2f} MiB", memory / static_cast<uint64_t>(1e6));
     }
-    else if(memory < 1e12)
+    else if(memory < static_cast<uint64_t>(1e12))
     {
-        return fmt::format("{:.2f} GiB", memory / 1e9);
+        return fmt::format("{:.2f} GiB", memory / static_cast<uint64_t>(1e9));
     }
-    return fmt::format("{:.2f} TiB", memory / 1e12);
+    return fmt::format("{:.2f} TiB", memory / static_cast<uint64_t>(1e12));
 }
